BELOW ERRORS FIXED
x^y not working. Check the image
https://imgur.com/EP6t3aO

Percent is not working
https://i.imgur.com/yKG6ETO.png

yrootx not working, also the button text is incorrect, y should be shown before x
https://i.imgur.com/y1aAJsh.png

If the operation buttons (x^2, rootx)are clicked without entering the number, it should be applied to 0

Inverse trigonometry not working for degree
eg. arcsin(1) should give output 90, if degree is selected

Please refer Google's scientific calculator for the functionalities